add test. change order
